requirements.txt
fastapi 
uvicorn
jinja2
python-multipart
mysql-connector-python